chore: fixed a few code quality issues reported by Qodana (in Intellij) (#819)
Fixed a few minor code quality issues reported by Qodana (in Intellij).

Solves PZ-1948

+++ src/main/java/net/atos/zac/app/contactmomenten/ContactmomentenRESTService.java
@@ -66,7 +66,7 @@ public class ContactmomentenRESTService {
         klantcontactmomentListParameters.setKlant(klant.getUrl());
         final var klantcontactmomentenResponse = contactmomentenClientService.listKlantcontactmomenten(klantcontactmomentListParameters);
         final List<RESTContactmoment> contactmomenten = klantcontactmomentenResponse.getResults().stream()
-                .skip(page * pageSize % NUM_ITEMS_PER_PAGE)
+                .skip((long) page * pageSize % NUM_ITEMS_PER_PAGE)
                 .limit(pageSize)
                 .map(this::convertToRESTContactmoment)
                 .toList();
